requirements.txt
websockets==16.0
pyyaml==6.0.3